Make directional light shadows more stable
Implement a technique to make directional light shadows more stable when
the camera moves around. The current swimming of the shadows is due to
the shadow bounds moving as the camera moves. Ensuring that the shadow
bounds are constrained to only move in texel-size increments (in
light/world space) ensures that the shadows are stable when the camera
moves around. There are still some artifacts when the camera rotates
though.